RBM Training works - (tested on one epoch)

+++ CBIRProjekt/CBIR_Project/src/iconn/htw/main/DefaultActivation.java
@@ -3,7 +3,7 @@ package iconn.htw.main;
 public class DefaultActivation implements IFloatFunction {
 
 	@Override
-	public float function(float x) {
+	public double function(double x) {
 		return x;
 	}
