+++ CBIRProjekt/CBIR_Project/src/iconn/htw/main/DefaultSigmoid.java
@@ -3,8 +3,8 @@ package iconn.htw.main;
 public class DefaultSigmoid implements IFloatFunction {
 
 	@Override
-	public float function(float x) {
-		return (float) (1.0 / (1.0 + Math.exp(-x)));
+	public double function(double x) {
+		return (double) (1.0 / (1.0 + Math.exp(-x)));
 	}
 
 }
