+++ CBIRProjekt/CBIR_Project/src/iconn/htw/main/IFloatFunction.java
@@ -2,5 +2,5 @@ package iconn.htw.main;
 
 public interface IFloatFunction {
 
-	float function(final float x);
+	double function(final double x);
 }
